Add fill_form_and_submit shortcut
The pattern of filling a form then immediately submitting seems to be
the most common interaction, and the more I use this the more I think
they should be combined.

I'm keeping both helpers separate as well in case there is something
that needs to be done between the interactions, so this is in addition
to the public interface.