Try translating option for select in StringInput
This will try translating the given option as defined by simple_form and
fallback to the actual value if no translation is found. If no option is
found matching the given value a more useful error will be raised
telling the user an option matching the given input was not found.