Add support for additional I18n label scopes
Prior to this commit, Formulaic assumed that all labels would be
declared as part of `simple_form` configuration.

This commit adds support for extending the list of possible
internationalization keys that Formulaic will look up.

For example, to lookup translations with [existing Rails
conventions][api-docs], users can configure `Formulaic` like so:

```rb
Formulaic.configure do |config|
  config.translation_scopes << "helpers.label"
  config.translation_scopes << "activerecord.attributes"
end
```

This is a looser follow-up pull request to [#67], which was denied and
closed on the grounds of keeping this gem SimpleForm-only.

While this change opens the door for users to extend the gem, it sticks
with SimpleForm defaults when un-configured.

[api-docs]: http://api.rubyonrails.org/classes/ActionView/Helpers/FormBuilder.html#method-i-label
[#67]: https://github.com/thoughtbot/formulaic/pull/67